chore(node/rpc): unify BaseFeeConfig to follow op-geth (#2198)
## Description

This PR unifies the `BaseFeeConfig` format between `op-geth` and the
`kona-node`.

`op-geth` format:
https://github.com/ethereum-optimism/op-geth/blob/6005dd53e1b50fe5a3f59764e3e2056a639eff2f/params/config.go#L509-L514
as they're using `u64` instead.

We were following `op-alloy`'s format until now:
https://github.com/alloy-rs/alloy/blob/0031ec3dd2ac6da303056374b86c1b4a8995bc34/crates/eips/src/eip1559/basefee.rs#L15-L26

+++ crates/node/engine/src/attributes.rs
@@ -898,8 +898,8 @@ mod tests {
     #[test]
     fn test_eip1559_default_param_cant_overflow() {
         let (mut cfg, mut attributes, mut block) = eip1559_test_setup();
-        cfg.chain_op_config.eip1559_denominator_canyon = u128::MAX;
-        cfg.chain_op_config.eip1559_elasticity = u128::MAX;
+        cfg.chain_op_config.eip1559_denominator_canyon = u64::MAX;
+        cfg.chain_op_config.eip1559_elasticity = u64::MAX;
 
         attributes.inner.eip_1559_params = Some(Default::default());
         block.header.extra_data = vec![0; 9].into();
@@ -912,8 +912,8 @@ mod tests {
             check,
             AttributesMatch::Mismatch(EIP1559Parameters(
                 BaseFeeParams {
-                    max_change_denominator: u128::MAX,
-                    elasticity_multiplier: u128::MAX
+                    max_change_denominator: u64::MAX as u128,
+                    elasticity_multiplier: u64::MAX as u128
                 },
                 BaseFeeParams { max_change_denominator: 0, elasticity_multiplier: 0 }
             ))
